app/requirements.txt
fastapi~=0.73.0
uvicorn~=0.15.0
psycopg~=3.1.10
psycopg-pool~=3.1.7
pydantic[dotenv]~=1.9.0
psycopg2-binary~=2.9.1
python-consul==1.1.0
